Optimize Content Cache Transmission with Scatter-Gather I/O
Accelerate sending content cache to the client by using
scatter-gather I/O.

The old version sends the content cache buffers to the client for
every linked list node. The new version uses scatter-gather I/O to
send the data to the client in a single transmission instead of
multiple transmissions.